+++ assets/javascripts/kitten/components/cards/crowdfunding-card/components/title.js
@@ -55,7 +55,7 @@ class TitleComponent extends PureComponent {
     titleTruncate: PropTypes.bool,
     loading: PropTypes.bool,
     widgetTitle: PropTypes.string,
-    dayCounter: PropTypes.string,
+    dayCounter: PropTypes.element,
     stateDay: PropTypes.string,
     titleProps: PropTypes.shape(),
   }
@@ -65,7 +65,7 @@ class TitleComponent extends PureComponent {
     titleTruncate: true,
     loading: false,
     widgetTitle: '',
-    dayCounter: '',
+    dayCounter: null,
     stateDay: '',
     titleProps: {},
   }