+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Delete `stateDay` and update `dayCounter` props on `CrowdfundingCardWidget`
+
 ## [2.37.2] - 2019-11-07
 
 Fix: